Split cars wizard lifecycle tests
