+++ apps/ui/tests/typecheck-baseline.txt
@@ -3,7 +3,7 @@
 tests/car_confidence_summary.spec.ts(134,5): error TS2353: Object literal may only specify known properties, and 'width_mm' does not exist in type '{ default_axle_for_speed: "front" | "rear" | "average"; front: { aspect_pct: number; rim_in: number; width_mm: number; }; name: string; rear?: { aspect_pct: number; rim_in: number; width_mm: number; } | null | undefined; rim_in: number; source_confidence?: string | ... 1 more ... | undefined; tire_aspect_pct: number...'.
 tests/car_confidence_summary.spec.ts(152,20): error TS2719: Type '{ default_axle_for_speed: "front" | "rear" | "average"; front: { aspect_pct: number; rim_in: number; width_mm: number; }; name: string; rear?: { aspect_pct: number; rim_in: number; width_mm: number; } | null | undefined; rim_in: number; source_confidence?: string | ... 1 more ... | undefined; tire_aspect_pct: number...' is not assignable to type '{ default_axle_for_speed: "front" | "rear" | "average"; front: { aspect_pct: number; rim_in: number; width_mm: number; }; name: string; rear?: { aspect_pct: number; rim_in: number; width_mm: number; } | null | undefined; rim_in: number; source_confidence?: string | ... 1 more ... | undefined; tire_aspect_pct: number...'. Two different types with this name exist, but they are unrelated.
 tests/car_wizard_view.spec.ts(65,3): error TS2322: Type '{ brand?: string | undefined; gearboxes: components["schemas"]["CarLibraryGearboxEntry"][]; model: string; rim_in: number; tire_aspect_pct: number; tire_options: components["schemas"]["CarLibraryTireOptionEntry"][]; tire_width_mm: number; type?: string | undefined; variants: components["schemas"]["CarLibraryVariantE...' is not assignable to type '{ brand: string; gearboxes: { final_drive_ratio: number; final_drive_ratio_confidence?: string | null | undefined; gear_ratios?: number[] | null | undefined; gear_ratios_confidence?: string | null | undefined; ... 5 more ...; transmission_confidence?: string | ... 1 more ... | undefined; }[]; ... 6 more ...; variant...'.
-tests/cars_feature_workflow.spec.ts(130,3): error TS2322: Type '{ brand?: string | undefined; gearboxes: components["schemas"]["CarLibraryGearboxEntry"][]; model: string; rim_in: number; tire_aspect_pct: number; tire_options: components["schemas"]["CarLibraryTireOptionEntry"][]; tire_width_mm: number; type?: string | undefined; variants: components["schemas"]["CarLibraryVariantE...' is not assignable to type '{ brand: string; gearboxes: { final_drive_ratio: number; final_drive_ratio_confidence?: string | null | undefined; gear_ratios?: number[] | null | undefined; gear_ratios_confidence?: string | null | undefined; ... 5 more ...; transmission_confidence?: string | ... 1 more ... | undefined; }[]; ... 6 more ...; variant...'.
+tests/cars_feature_workflow.spec.ts(106,3): error TS2322: Type '{ brand?: string | undefined; gearboxes: components["schemas"]["CarLibraryGearboxEntry"][]; model: string; rim_in: number; tire_aspect_pct: number; tire_options: components["schemas"]["CarLibraryTireOptionEntry"][]; tire_width_mm: number; type?: string | undefined; variants: components["schemas"]["CarLibraryVariantE...' is not assignable to type '{ brand: string; gearboxes: { final_drive_ratio: number; final_drive_ratio_confidence?: string | null | undefined; gear_ratios?: number[] | null | undefined; gear_ratios_confidence?: string | null | undefined; ... 5 more ...; transmission_confidence?: string | ... 1 more ... | undefined; }[]; ... 6 more ...; variant...'.
 tests/dom_render_test_support.ts(121,7): error TS2352: Conversion of type '{ getPropertyValue: () => string; }' to type 'CSSStyleDeclaration' may be a mistake because neither type sufficiently overlaps with the other. If this was intentional, convert the expression to 'unknown' first.
 tests/dom_render_test_support.ts(160,5): error TS2322: Type 'NonNullable<MountedViewGlobals[K]>' is not assignable to type '(typeof globalThis & MountedViewGlobals)[K]'.
 tests/dom_render_test_support.ts(55,67): error TS2345: Argument of type 'TView | undefined' is not assignable to parameter of type 'TView'.
